+++ trunk/l2jfree-core/src/main/java/com/l2jfree/gameserver/model/actor/instance/L2CastleChamberlainInstance.java
@@ -37,6 +37,7 @@ import com.l2jfree.gameserver.model.L2Clan;
 import com.l2jfree.gameserver.model.L2Skill;
 import com.l2jfree.gameserver.model.L2TeleportLocation;
 import com.l2jfree.gameserver.model.entity.Castle;
+import com.l2jfree.gameserver.model.itemcontainer.PcInventory;
 import com.l2jfree.gameserver.network.SystemMessageId;
 import com.l2jfree.gameserver.network.serverpackets.ActionFailed;
 import com.l2jfree.gameserver.network.serverpackets.ExShowCropInfo;
@@ -262,17 +263,17 @@ public class L2CastleChamberlainInstance extends L2MerchantInstance
 					return;
 
 				String filename = "data/html/chamberlain/chamberlain-vault.htm";
-				int amount = 0;
+				long amount = 0;
 				if (val.equals("deposit"))
 				{
 					try
 					{
-						amount = Integer.parseInt(st.nextToken());
+						amount = Long.parseLong(st.nextToken());
 					}
 					catch (NoSuchElementException e)
 					{
 					}
-					if (amount > 0 && (long) getCastle().getTreasury() + amount < Integer.MAX_VALUE)
+					if (amount > 0 && getCastle().getTreasury() + amount < Integer.MAX_VALUE)
 					{
 						if (player.reduceAdena("Castle", amount, this, true))
 							getCastle().addToTreasuryNoTax(amount);
@@ -284,7 +285,7 @@ public class L2CastleChamberlainInstance extends L2MerchantInstance
 				{
 					try
 					{
-						amount = Integer.parseInt(st.nextToken());
+						amount = Long.parseLong(st.nextToken());
 					}
 					catch (NoSuchElementException e)
 					{
